Refactor configuration and HTML generation into utility classes.
